pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg==0.29.0",
    "black==23.11.0",
    "celery==5.3.4",
    "fastapi==0.104.1",
    "flake8==6.1.0",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "httpx==0.25.2",
    "librosa==0.10.1",
    "mutagen==1.47.0",
    "numpy==1.24.3",
    "oauthlib>=3.3.1",
    "passlib[bcrypt]==1.7.4",
    "pathlib2==2.3.7",
    "pydantic[email]==2.11.7",
    "pydub==0.25.1",
    "pyjwt>=2.10.1",
    "pytest==7.4.3",
    "python-jose[cryptography]==3.3.0",
    "python-magic==0.4.27",
    "python-multipart==0.0.6",
    "redis==5.0.1",
    "scipy==1.11.4",
    "soundfile==0.12.1",
    "uvicorn[standard]==0.24.0",
    "websockets==12.0",
]
